requirements.txt
streamlit
pdfplumber
PyPDF2
python-docx
